Fixup

+++ spring-aop/src/test/java/org/springframework/aop/aspectj/AspectJAdviceParameterNameDiscovererTests.java
@@ -235,7 +235,7 @@ public class AspectJAdviceParameterNameDiscovererTests {
 	protected void assertParameterNames(
 			Method method, String pointcut, String returning, String throwing, String[] parameterNames) {
 
-		assertThat(parameterNames.length).as("bad test specification, must have same number of parameter names as method arguments").isEqualTo((long) method.getParameterCount());
+		assertThat(parameterNames.length).as("bad test specification, must have same number of parameter names as method arguments").isEqualTo(method.getParameterCount());
 
 		AspectJAdviceParameterNameDiscoverer discoverer = new AspectJAdviceParameterNameDiscoverer(pointcut);
 		discoverer.setRaiseExceptions(true);
@@ -248,7 +248,7 @@ public class AspectJAdviceParameterNameDiscovererTests {
 
 		assertThat(discoveredNames.length).as("Expecting " + parameterNames.length + " parameter names in return set '" +
 				formattedExpectedNames + "', but found " + discoveredNames.length +
-				" '" + formattedActualNames + "'").isEqualTo((long) parameterNames.length);
+				" '" + formattedActualNames + "'").isEqualTo(parameterNames.length);
 
 		for (int i = 0; i < discoveredNames.length; i++) {
 			assertThat(discoveredNames[i]).as("Parameter names must never be null").isNotNull();
